Added btrfs option

+++ archmatic.sh
@@ -132,6 +132,9 @@ function preinstall {
     echo -e "\nSelect your CPU: [amd/intel/vbox]"
     read ucode
 
+    echo -e "\nSelect desired filesystem: [ext4/btrfs]"
+    read filesystem
+
     # export environment variabels
     export disk
     export disk_boot
@@ -148,6 +151,7 @@ function preinstall {
     export bluetooth
     export wifi
     export ucode
+    export filesystem
 }
 
 function baseInstall {
@@ -191,7 +195,18 @@ function baseInstall {
     # Create filesystem
     mkfs.ext4 ${disk_boot}
     mkfs.fat -F32 ${disk_esp}
-    mkfs.ext4 /dev/vg0/lv_root
+
+    ## Filesystem for root-volume
+    
+    ### EXT4
+    if [ "${filesystem}" == "ext4" ]; then
+        mkfs.ext4 /dev/vg0/lv_root
+    fi
+
+    ### BTRFS
+    if [ "${filesystem}" == "btrfs" ]; then
+        mkfs.btrfs /dev/vg0/lv_root
+    fi
 
     # Mount target
     mount /dev/vg0/lv_root /mnt
